p2v: Fix wait_qemu_nbd to use getaddrinfo.
wait_qemu_nbd would fail in Rawhide where /etc/hosts contains:

  127.0.0.1   localhost [...]
  ::1         localhost [...]

This peculiar mapping caused qemu-nbd to always select an IPv6
address.  However since we used the obsolete inet_pton function and
forced AF_INET, we were always trying to connect to the IPv4
localhost.

As this was never going to work anyway, fix the code to use
getaddrinfo to resolve "localhost", and use a loop to connect to all
possible addresses.

Also: Add a short sleep in the reconnection loop so that we don't
hammer the qemu-nbd server while it is starting up.