p2v: Fix parsing of p2v.memory parameter (RHBZ#1229262).
The sscanf call always failed because we checked for the wrong return
code.

We don't need to allocate two characters for the mem_code field since
sscanf doesn't append a nul byte.

This commit also allows you to use lowercase 'm' or 'g' as a memory
unit.

Also clarify the documentation: the memory unit must be given.