p2v: Pass host CPU details to virt-v2v.
In the fake <domain type='physical'> libvirt XML that we create to
describe the physical host, we did not accurately pass any information
about the host CPU except the number of cores (<vcpu/>).

This commit extracts detailed information about the vendor, model and
topology of the host CPU and adds that to the libvirt XML for
virt-v2v.  Conveniently we can use libvirt capabilities to get this
information without needing to parse /proc/cpuinfo or similar
techniques.

The libvirt XML looks like this:

  <domain type="physical">
  ...
    <cpu match="minimum">
      <vendor>Intel</vendor>
      <model fallback="allow">Broadwell</model>
      <topology sockets="1" cores="2" threads="2"/>
    </cpu>
  ...
    <features>
      <acpi/>
      <apic/>
      <pae/>
    </features>