make-disk: set p2v.local as hostname
Set p2v.local as hostname for the p2v disk image, so it will be fixed,
instead of whatever is set in the used templates.  This is also somewhat
important during the distribution update, and installation of our
packages: there may be distribution packages actually using the hostname
to initialize stuff (e.g. configuration values), and thus the unset
value "(none)" triggers errors.

The additional command to set the hostname is for the appliance itself:
the hostname will be set on the running kernel, which is the
appliance's own, so getting it later on with hostname(1) works.