Use g_spawn_sync to launch curl
Instead of using system() with stderr redirected to file, use the GLib
function g_spawn_sync() to run curl with no need for a shell, and
directly collecting stderr as a string.

The only behaviour change is that a failure when spawning curl is now
logged as error, instead of aborting the whole p2v.