make-disk: Don't upgrade the kernel to avoid kernel command line bug
There is some bug in recent Fedora which prevents the kernel from
being upgraded correctly from libguestfs (eg. using
virt-customize --update or virt-builder --update).  As an example of
what happens and how to reproduce it see:

https://bugzilla.redhat.com/show_bug.cgi?id=1945835#c24

This problem causes the kernel command line to be messed up, adding
bits of the libguestfs appliance /proc/cmdline including a bogus root=
parameter, and this prevents virt-p2v from finding the root filesystem
when booting.

This patch is just a workaround for this bug.  By preventing the
kernel* package from being upgraded we continue to use the working
kernel from the template.  Userspace packages in the template get
updated as before.

Acked-by: Laszlo Ersek <lersek@redhat.com>