do not connect to nbdkit after starting it
Consider the call tree again:

  start_conversion()                [conversion.c]
    start_nbd_server()              [nbd.c]
      open_listening_socket()       [nbd.c]
        bind_tcpip_socket()         [nbd.c]
          socket()
          bind()
          listen()
      start_nbdkit()                [nbd.c]
        fork()
          /* child */
          execlp()
      close()
    wait_for_nbd_server_to_start()  [nbd.c]
      connect_to_nbdkit()           [nbd.c]
        connect()
    open_data_connection()          [ssh.c]
      /* "-R 0:localhost:<port>" */

Right after we fork(), inside (indirectly) start_nbd_server(), two file
descriptors exist that point to the same file description that points to
the same listening socket. In the child, execlp() preserves one of the
file descriptors (that's the point of socket activation), while in the
rest of start_nbd_server(), the parent close()s the other file descriptor.

By the time the parent reaches connect(), there will have been no
opportunity for the listening socket not to exist (unless, in the child,
execlp() fails, or the new image exits very quickly). Therefore we don't
need to test the child (by connecting to it and reading NBDMAGIC); we can
just expose the listening socket to "ssh -R".

Thus, remove the wait_for_nbd_server_to_start() and connect_to_nbdkit()
functions.

This eliminates the following -- harmless -- message during conversion,
emitted by the nbdkit server child on the p2v machine, to the serial port:

> nbdkit: file[1]: error: reading initial client flags: conn->recv:
> Connection reset by peer

nbdkit complains because, after sending "NBDMAGIC",  it would like to send
"IHAVEOPT" and the handshake flags as well to the client. But the client
-- namely wait_for_nbd_server_to_start() -- stops reading from the socket
right after seeing "NBDMAGIC" (8 bytes). So that's a "broken pipe"
situation (RST) for the sending party (the server).

Ref: https://listman.redhat.com/archives/libguestfs/2022-March/028528.html
Suggested-by: Richard W.M. Jones <rjones@redhat.com>
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220401102808.11062-12-lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>