virt-p2v-make-*.in: make @datadir@ and @libdir@ resolvable
./configure expands @datadir@ and @libdir@ to "${prefix}/share" and
"${exec_prefix}/lib" by default (verbatim); $prefix and $exec_prefix are
supposed to be substituted by the shell when the generated shell scripts
are executed.

For this, capture the prefix and exec_prefix values at the time
./configure runs.

This fixes failures such as

> $ virt-p2v-make-kickstart fedora
> base64: /share/virt-p2v/issue: No such file or directory

for such scripts that were installed with "make install".

Note that the patch does not help when running virt-p2v-make-* from a git
checkout that was just built, such as

> $ ./virt-p2v-make-kickstart fedora

In that case, VIRT_P2V_DATA_DIR still needs to be set to $PWD, as there is
no "share" subdirectory in the checkout.

Edited to capture datarootdir too, at Eric's suggestion, "as it will then
silence the autoconf warnings about not honoring --datarootdir in a file
that dereferenced @datadir@".

Suggested-by: Richard W.M. Jones <rjones@redhat.com>
Suggested-by: Eric Blake <eblake@redhat.com>
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220720115322.17665-1-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>
Acked-by: Eric Blake <eblake@redhat.com>