Makefile.am: factor out "make-physical-machine.sh"
Extract and somewhat generalize the recipe for the $(PHYSICAL_MACHINE)
target to a separate shell script. In preparation for the multiple steps
we're going to introduce later, redirect virt-builder to a temp file at
first (placed in the same directory as the finally expected disk image),
and rename that file upon success.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220826113908.12542-2-lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>