gui: offer copying the vCPU topology from the fully populated physical one
Introduce a checkbox that lets the user choose between options (a) and
(b), as outlined in one of the previous patches in this series. When the
checkbox is ticked -- that is, the fully populated physical topology is
copied --, make the VCPU count text entry insensitive, and overwrite its
contents with the value calculated from the physical topology. Upon
unticking the checkbox, restore the last manual entry.

The VCPU limit warning applies automatically when the checkbox is toggled,
because when the VCPU count text entry is programmatically modified, the
entry emits the "changed" signal, and that chains to
vcpus_or_memory_check_callback().

However, in show_conversion_dialog(), we need to handle the case of option
(a) potentially becoming the default. In that case we have to render the
VCPU count text entry insensitive immediately, and overwrite its contents
(set originally from the *online* logical processor count, or on the
kernel command line) with the full population count derived from the
topology. For this, call vcpu_topo_toggled() manually. This in turn may
(as it should) trigger the VCPU limit warning at once, handled by the
vcpus_or_memory_check_callback() call right after.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1590721
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>
Message-Id: <20220908162746.25031-5-lersek@redhat.com>