eliminate shadowing of global variables
Shadowing of global variables, especially if it's intentional and/or as
wide-spread in a codebase as it is in virt-p2v, is very harmful. Eliminate
it by renaming a number of function parameters, appending a "_p" suffix.

(The shadowing of "v2v_version" interfered with the analysis of the crash
that I originally meant to fix today. The fix is now the last patch in the
series.)

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220911142501.18230-3-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>