ssh.c: fix crash in "v2v_version" lifecycle management
We only ever set "v2v_version" to non-NULL with
pcre2_substring_get_bynumber(), as suggested by valgrind and then verified
by code review.

Substrings extracted with this function cannot be released with free(); we
must call pcre2_substring_free().

http://www.pcre.org/current/doc/html/pcre2_substring_get_bynumber.html
http://www.pcre.org/current/doc/html/pcre2_substring_free.html

The symptom of this bug is that the second click on the "Test Connection"
causes free() to print:

> free(): invalid pointer

and then to call abort().

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220911142501.18230-5-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>