gui.c: consume "all_disks" and "all_removables" only for dialog setup
Currently "all_disks" and "all_removables" are used for both populating
the conversion dialog and collecting the configuration updates from the
conversion dialog.

Going forward, we want to eliminate "all_disks" and "all_removables" as
global variables. As one step in that direction, restrict the consumption
of "all_disks" and "all_removables" to when we populate the conversion
dialog. For this, introduce a new column to the list stores that back the
"fixed disks" and "removable media drives" tree views, and copy
"all_disks" and "all_removables" into the new columns, respectively. This
new column is never displayed; it's only used as a lookaside buffer when
we collect the disks / removables selections into the config structure.

Note that both gtk_list_store_set() and gtk_tree_model_get() create deep
copies of strings.

No observable changes.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2124538
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220919133511.18288-4-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>