create_conversion_dialog(): add button to refresh disks & removables
Using the previously extracted / reworked representation of disks and
removables, introduce a new button in the conversion dialog that
repopulates the "disks" and "removables" tree views with freshly collected
entries. In case the user issues such commands in the XTerm window that
make more disks visible to the kernel, for example iSCSI LUNs, the button
allows virt-p2v to learn about those disks.

Disable the new button for when "--test-disk" is passed on the command
line; that option conflicts with the new feature. (This means the new
button is insensitive in "make run-virt-p2v-directly". It is sensitive in
the VM tests though.)

Update the manual accordingly.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2124538
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220919133511.18288-15-lersek@redhat.com>
[lersek@redhat.com: refresh the disk/network configuration dialog's width
 in the manual; also replacing an incorrectly used pipe sign (U+007C,
 Vertical Line) with U+2502 (Box Drawings Light Vertical) ]
Acked-by: Richard W.M. Jones <rjones@redhat.com>