gui.c: remove line truncation (originally a workaround for a GTK2 bug)
Commit eaa5ce8b43e0 ("p2v: Break long lines when displaying virt-v2v
output in GUI.", 2016-01-08) introduced *line breaking* as a workaround
for a GTK2 regression, under which the handling of long lines from the
virt-v2v log would slow down very much.

Commit e09b955bdbbc ("p2v: Use virt-v2v --colours option, support colour
in the run dialog (RHBZ#1314244).", 2016-06-18) silently changed the *line
breaking* workaround into *line truncation*, also adopting a much shorter
cut-off length (replacing 1024 with 256).

Now, whether this logic remains useful is untestable today.  That's
because commit 9b057cc1ac5f ("p2v: Don't display debugging messages in the
run dialog.", 2016-06-18), from the same series as the one that commit
e09b955bdbbc belonged to [*], had relegated the verbose & trace messages
-- IOW, all the long lines -- from the GUI to the conversion log file
(except maybe the last 50 lines of a failed conversion's log).

[*] https://bugzilla.redhat.com/show_bug.cgi?id=1314244#c4

Therefore, the line truncation workaround may have become effectively
useless right at the end of that series, and stays most probably dead code
today, regardless of the GTK version used.

This is relevant because we want to remove GTK2 support from virt-p2v.
Some evidence would be nice as to whether the line truncation is obsolete
due to (a) GTK3 not suffering from the long line slowdown, or (b) GTK3
still having the bug, but the bug not being triggered due to commit
9b057cc1ac5f.  Either way -- that is, even though this is untestable --,
just remove the truncating state from the v2v message parsing automaton.

The truncating state is only reachable from the normal state, and it also
returns to the normal state, so the removal is not complicated.  Remove
the "dots" helper variable too.  The (static) variable "linelen" is only
consumed in the normal->truncating transition condition, so drop that one
as well.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220926081812.12585-5-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>