gui.c: fix warning label wrapping regression under GTK3
(Note that this issue is reproducible with virt-p2v v1.42.2.)

When building virt-p2v with GTK3 (specifically,
gtk3-3.24.34-1.fc35.x86_64), the label that warns about too many VCPUs
and/or too much guest RAM does not wrap, even though we set it to
wrapping.  Instead, the (otherwise fixed size) outer window is resized
with a sudden jump, to accommodate the (now very long) single-line
warning.

This issue does not manifest itself under GTK2, which is why I've not
noticed it thus far.  Under GTK3, we can mitigate the symptom by following
the advice at:

  https://wiki.gnome.org/HowDoI/Labels#Labels_in_non-resizable_windows

Namely, set the "max-width-chars" property of the label, specifying the
"natural width" of the label, expressed in units of average-width
characters.  The empirical value 50 seems to do what GTK2 does out of the
box.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220926081812.12585-6-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>