Makefile.am: rectify target dependencies on package dependency files
The targets whose recipes invoke "virt-p2v-make-disk" should depend on
"$(dependencies_files)", and not "dependencies.m4". The reason is that
"virt-p2v-make-disk" consumes the filtered / processed files, not the
original dependency list. We have a separate rule that invokes m4 for
generating "$(dependencies_files)" from "dependencies.m4"; we need to loop
that in, so that "virt-p2v-make-disk" see up-to-date filtered files.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220930145252.16273-2-lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>