tests: Allow virt-builder osversion to be overridden
When testing on Fedora 37 (which has not yet been released):

  $ make run-virt-p2v-in-a-vm
  [...]
  ./run virt-p2v-make-disk -o virt-p2v.img
  virt-builder: error: cannot find os-version ‘fedora-37’ with
  architecture ‘x86_64’.
  Use --list to list available guest types.

Allow osversion to be overridden in this case.  It defaults to empty
which means, as before, select the current host osversion.

  $ make run-virt-p2v-in-a-vm VIRT_P2V_OSVERSION=fedora-36

Note that the osversion must still be compatible with the binary.
eg. This cannot be used to test virt-p2v on RHEL 7 on a Fedora host
because the binary built on Fedora would not run on RHEL 7, so the
utility of this is quite narrow.  (Hence why I didn't bother to
document it except as a comment.)