make-disk: restrict kernel update bug work-around
Commit 2764abefaa14 ("make-disk: Don't upgrade the kernel to avoid kernel
command line bug", 2022-03-24) replaced the "--update" option of
virt-builder with an open-coded DNF update command, excluding the kernel
packages from the update.

Albeit a justified workaround, it is a bit too broad: it breaks on all
guest OS images that don't have DNF, such as old RHEL, Debian and Ubuntu,
etc.  (In a subsequent patch, I'm going to refactor the Debian/Ubuntu
branch, and for testing that, I first need eliminate the open-coded DNF
command.)  Refine the workaround as follows:

- I've tested Fedora 33 and Fedora 37; they work without the workaround
  (they are not yet, and no longer, affected by the bug, respectively).

- I've tested Fedora 34 and 36, they are affected.  They need the
  workaround.

- Fedora 35 is also broken, per
  <https://bugzilla.redhat.com/show_bug.cgi?id=1945835#c24>.

- RHEL-9.0 and RHEL-9.1 *seem* to work fine without the workaround (that
  is, with commit 2764abefaa14 reverted).  However, that's deceptive: in
  my testing, the original "--update" option does not bring in any new
  kernel packages, so even if the bug is still there, it is not triggered.
  Because RHEL-9 is based on Fedora 34, which is affected, assume that
  RHEL-9 is affected too.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20230119121448.107936-3-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>