make-disk: replace here-docs with printfs for better indentation
Here-documents break the indentation of the script; replace them with the
"printf" shell utility. (We could use the <<- redirection operator for
stripping leading TABs from the here-docs, but the script itself does not
use TABs for indentation, so that would introduce a different kind of
inconsistency.)

Because we use a quoted word ('EOF') for our delimiter in the
here-documents, variables in the here-docs are not expanded; stick with
that.

Use a sole '%s\n' format operand with each printf invocation: "[t]he
/format/ operand shall be reused as often as necessary to satisfy the
/argument/ operands"
<https://pubs.opengroup.org/onlinepubs/9699919799/utilities/printf.html>.

(I've tried to regression-test this with the two latest Ubuntu images:
ubuntu-18.04, ubuntu-20.04. The former wouldn't build, and the latter only
booted to a GDM login screen. My build host is RHEL-9.1, so the kiosk is
not actually expected to work. Any further investigation here is left to
Ubuntu users.

The Fedora 37 image works fine, even with the virt-p2v binary coming from
RHEL-9.1.)

Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20230119121448.107936-4-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>